don't trigger push if already up to date
This was causing a problem with bitbucket. All repositories were "last updated" just after git2r::push() even if the ref has not changed. The problem did not occur when pusing from the command line. Also, it is nice to see in the console whether or not a remote repo got updated.